iot-final-project/python/requirements.txt
paho-mqtt
matplotlib
